floppy.c
- floppyHasIndex() tested motor_tick == 0, when motor_tick was
oddly aligned, this test would never trigger since the code reads 2 bytes
forward every time. The loading then halted in Beast 3.

- The disksync value in the mfm_data was not word aligned, so the code never
saw it. I added a temporary fix in floppyEndOfLineC() that steps one
byte back to synchronize the mfm-stream to it.
A more comfortable solution is to rewrite the code that reads word-wise to read bytes
into a word-window instead.

caps_win32.c

In capsLoadImage(), there was a minor error in the tracks calculation
that caused floppy[].tracks to be doubled. (84 vs 168). It allowed
Beast 3 to step outside the initialized tracks.