sprite.c/.h
Optimized the use of action lists to get rid of mallocs.
Removed the use of list.h for managing the sprite action lists.
It is now using pre-allocated arrays of action item structs.
Added functions needed to manage the lists, and changed the
code that used them. This resulted in a huge speed increase.