Added all clip config settings. +ini file, change window pos to signed.
Clip mode "fixed" keeps the current values set in code. For now the GUI
will use fixed clip and set the clip to max PAL values in automatic
scale mode. Must find out how to present this in a UI. Also a problem
with the window starting outside the desktop might be fixed by using
signed coordinates in the ini-file.